+++ WerkWerk/WorkContext.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Threading;
 using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 using System.Text.Json;
@@ -9,6 +10,7 @@ namespace WerkWerk
 
     public class WorkContext<T>
     {
+        private CancellationToken _jobCancel;
         private ILogger _logger;
         private IServiceProvider _services;
         private Dictionary<string, object> _items = new Dictionary<string, object>();
@@ -17,15 +19,16 @@ namespace WerkWerk
         public T Data { get; private set; }
         public ILogger Logger => _logger;
         public IDictionary<string, object> Items => _items;
-
+        public CancellationToken JobCancel => _jobCancel;
         public IServiceProvider Services => _services;
 
-        internal static WorkContext<T> FromJob(Job job, ILogger logger, IServiceProvider services)
+        internal static WorkContext<T> FromJob(Job job, ILogger logger, IServiceProvider services, CancellationToken token = default)
         {
             return new WorkContext<T>
             {
                 _logger = logger,
                 _services = services,
+                _jobCancel = token,
                 RequestedBy = job.RequestedBy,
                 Data = JsonSerializer.Deserialize<T>(job.Data),
             };